# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "description": "Admin: todos (filtro opcional companyId). Business: solo su empresa.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "422": {
            "description": "Business sin empresa configurada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar usuarios",
        "tags": [
          "Usuarios (Admin)"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualizar usuario",
        "tags": [
          "Usuarios (Admin)"
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "description": "Alterna el campo isActive del usuario.",
        "operationId": "UsersController_toggleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "El actor intenta deshabilitarse a sí mismo"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Habilitar o deshabilitar usuario",
        "tags": [
          "Usuarios (Admin)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cotizador API",
    "description": "Contrato OpenAPI 3 para el front (`cotizador-app`).\n\nPrefijo global: `/api`. Auth: header `Authorization: Bearer <jwt>` (scheme `access-token`).\nJSON vivo: `GET /api/openapi.json`. UI: `/api/docs`.\nGuía de pantallas, roles y labels: `docs/contrato-front.md`.\n\nErrores: `{ statusCode, message, traceId? }`. `message` puede ser `string` o `string[]` (validación).\nMontos en CLP (número). Fechas de negocio `YYYY-MM-DD`; timestamps ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token obtenido en POST /api/auth/login"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "mobilePhone": {
            "type": "string",
            "example": "+56 9 1234 5678"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "business",
              "common"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "mobilePhone",
          "role",
          "isActive",
          "mustChangePassword",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nombre actualizado"
          },
          "mobilePhone": {
            "type": "string",
            "example": "+56 9 9999 8888"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "business",
              "common"
            ]
          }
        }
      },
...
    }
  }
}